+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
+## v0.71.7 | 24 Mar 2024
+
+- Refactored code for class `HeatmapDateTime` (`diive.core.plotting.heatmap_datetime.HeatmapDateTime`)
+- Fixed bug in `HeatmapDateTime` where the last record of each day was not shown
+
 ## v0.71.6 | 23 Mar 2024
 
 ![DIIVE](images/analysesZaggregatesInQuantileClassesOfXY_diive_v0.71.6.png)